// client-app/src/app/models/products.ts
export interface Product {
  id: string
  name: string
  description: string
  imageSrc: string
  price: number
  department: string
  category: string
  supplierProductId: number
  hasDiscount: boolean
  discountValue: number
  gallery: any
  material: string
  details: any
  quantity: number
  total: number
  uniqueId: string
}